+++ imglib2/core/src/test/java/net/imglib2/view/RandomAccessTest.java
@@ -2,10 +2,11 @@
  * #%L
  * ImgLib2: a general-purpose, multidimensional image processing library.
  * %%
- * Copyright (C) 2009 - 2012 Stephan Preibisch, Stephan Saalfeld, Tobias
- * Pietzsch, Albert Cardona, Barry DeZonia, Curtis Rueden, Lee Kamentsky, Larry
- * Lindsey, Johannes Schindelin, Christian Dietz, Grant Harris, Jean-Yves
- * Tinevez, Steffen Jaensch, Mark Longair, Nick Perry, and Jan Funke.
+ * Copyright (C) 2009 - 2013 Stephan Preibisch, Tobias Pietzsch, Barry DeZonia,
+ * Stephan Saalfeld, Albert Cardona, Curtis Rueden, Christian Dietz, Jean-Yves
+ * Tinevez, Johannes Schindelin, Lee Kamentsky, Larry Lindsey, Grant Harris,
+ * Mark Hiner, Aivar Grislis, Martin Horn, Nick Perry, Michael Zinsmaier,
+ * Steffen Jaensch, Jan Funke, Mark Longair, and Dimiter Prodanov.
  * %%
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions are met:
@@ -37,7 +38,6 @@
 package net.imglib2.view;
 
 import static org.junit.Assert.assertTrue;
-
 import net.imglib2.RandomAccess;
 import net.imglib2.img.Img;
 import net.imglib2.img.array.ArrayImgFactory;
@@ -109,7 +109,7 @@ public class RandomAccessTest
 		for( int d = 0; d < a.numDimensions(); ++d )
 		{
 			Assert.assertTrue( expected[ d ] == a.getLongPosition( d ) );
-			Assert.assertTrue( expected[ d ] == (long) a.getIntPosition( d ) );
+			Assert.assertTrue( expected[ d ] == a.getIntPosition( d ) );
 			Assert.assertTrue( expected[ d ] == (long) a.getFloatPosition( d ) );
 			Assert.assertTrue( expected[ d ] == (long) a.getDoublePosition( d ) );
 		}
